move deck initialize process
